requirements.txt
blinker==1.9.0
click==8.3.1
filelock==3.24.2
Flask==3.1.2
flask-cors==6.0.2
fsspec==2026.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
pillow==12.1.1
setuptools==82.0.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
Werkzeug==3.1.5
